x671 - visiumQCSpec change
